Add support for multiple profiles

+++ src/seeds/009_redirect.js
@@ -1,19 +1,27 @@
 import { map } from 'lodash';
 
-import { log, stripI18n } from '../helpers';
+import { fileExists, log, stripI18n } from '../helpers';
 import { Redirect } from '../models';
+
+const { config } = require(`${process.cwd()}/config`);
 
 export const seed = async (knex) => {
   try {
-    const profile = stripI18n(require('../profiles/redirects'));
-    if (profile.purge) {
-      await Redirect.delete(knex);
-    }
     await Promise.all(
-      map(
-        profile.redirects,
-        async (redirect) => await Redirect.create(redirects, {}, knex),
-      ),
+      map(config.profiles, async (profilePath) => {
+        if (fileExists(`${profilePath}/redirects`)) {
+          const profile = stripI18n(require(`${profilePath}/redirects`));
+          if (profile.purge) {
+            await Redirect.delete(knex);
+          }
+          await Promise.all(
+            map(
+              profile.redirects,
+              async (redirect) => await Redirect.create(redirects, {}, knex),
+            ),
+          );
+        }
+      }),
     );
     log.info('Redirects imported');
   } catch (err) {
